+++ src/main/java/cat/soft/src/parking/model/room/GetJoinRoomReq.java
@@ -14,5 +14,5 @@ public class GetJoinRoomReq {
 	@NotNull(message = "방 번호를 입력하세요.")
 	@Min(value = 0, message = "방 번호를 확인하세요")
 	@Max(value = Integer.MAX_VALUE, message = "방 번호를 확인하세요.")
-	private Integer room_id;
+	private Long room_id;
 }
