+++ src/main/java/cat/soft/src/parking/repository/TimeRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cat.soft.src.parking.model.Time;
 
-public interface TimeRepository extends JpaRepository<Time, Integer> {
+public interface TimeRepository extends JpaRepository<Time, Long> {
 
-	Time findTimeByParkingLotIdxAndRoomIdxAndEndAfter(Integer parkingLotIdx, Integer roomIdx, ZonedDateTime time);
+	Time findTimeByParkingLotIdxAndRoomIdxAndEndAfter(Long parkingLotIdx, Long roomIdx, ZonedDateTime time);
 
-	Time findTimeByUserIdxAndEndAfter(Integer userIdx, ZonedDateTime time);
+	Time findTimeByUserIdxAndEndAfter(Long userIdx, ZonedDateTime time);
 }
